backend/pyproject.toml
[project]
name = "awsrt-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "pydantic>=2",
  "numpy",
  "pandas",
  "pyarrow",
  "duckdb",
  "zarr",
  "sse-starlette",
  "websockets",
  "python-multipart",
  "rich",
  "typer",
  "numcodecs",
  "matplotlib",
  "pillow",
]

[tool.uvicorn]
factory = false
host = "0.0.0.0"
port = 8000
reload = true
